# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.14"
  title: DTL Service
# during dev, should point to your local machine
host: localhost:80
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/octet-stream #smell, this is necessary because supertest appears to send this content type for empty posts
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/plain
  - text/html
paths:
...
  /findByGeoLocation/venue/{venueId}:
    x-swagger-router-controller: WpGeoLocationController
    get:
      description: Get Experience Structure by Experience ID
      operationId: findByGeoLocation
      parameters:
      - name: venueId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Venue
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EntityResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"  
  /findByGeoLocation/query/{entityName}:
    x-swagger-router-controller: WpGeoLocationController
    post:
      description: Get Experience Structure by Experience ID
      operationId: findByGeoLocationPost
      parameters:
      - name: entityName
        type: string
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      - name: body
        in: body
        required: true
        description: JSON document with geo query parameters 
        schema:
          type: object
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EntityResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"  
...
definitions:
...
  EntityResponse:
    type: object
  # Response definitions
  ErrorResponse:
    required:
      - type
      - code
      - message
    properties:
      type:
        type: string
      code:
        type: number
      message:
        type: string
...